requirements.txt
pywebview>=4.4.1
pyinstaller>=6.0.0

